backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
pydantic>=2.9,<3.0
pydantic-settings>=2.6,<3.0
python-dotenv>=1.0,<2.0
openai>=1.54,<2.0
pyjwt[crypto]>=2.9,<3.0
supabase>=2.9,<3.0
httpx>=0.27,<1.0
